implemented authenticate + added gig routes
implemented authenticate + getNearbyGigs  + getRecommendedGigs +
calculation for distance util

+++ src/main/kotlin/com/swiftshift/data/model/Gig.kt
@@ -14,8 +14,8 @@ data class Gig(
     val currentApplier: Int,
     val deadlineDate: Long,
     val salary: Double,
-    val latitude: Long,
-    val longitude: Long,
+    val latitude: Double,
+    val longitude: Double,
     val timestamp: Long,
     @BsonId
     val id: String = ObjectId().toString()
